Refactor into package with blueprints
